Move custom components for ReactSelect into separate files
