+++ src/react/src/components/Filters/NestedSelect/CustomDropdownIndicator.jsx
@@ -1,0 +1,10 @@
+import React from 'react';
+import ArrowDropDownIcon from '../../ArrowDropDownIcon';
+
+const CustomDropdownIndicator = ({ selectProps: { classes } }) => (
+    <div className={classes.dropdownIndicator}>
+        <ArrowDropDownIcon />
+    </div>
+);
+
+export default CustomDropdownIndicator;
